chore: Update main.dart with color scheme change
The main.dart file has been updated to change the color scheme used in the application. The previous color scheme was based on the deepPurple color, but it has been changed to use the greenAccent color. This change aims to improve the visual appearance of the application and provide a more cohesive design.