chore: Update main.dart with lazy loading for book details page
The main.dart file has been updated to implement lazy loading for the book details page. The '/detail-book' route has been commented out to prevent unnecessary loading of the BookPage component. This change aims to optimize the performance of the application by only loading the book details page when necessary.